Refactor React folder structure
