+++ hanki-frontend/src/App.js
@@ -1,5 +1,5 @@
 import "./App.css";
-import Home from "./components/Home";
+import Home from "./layout/Home";
 import "@fortawesome/fontawesome-free/css/all.min.css";
 
 function App() {
